feat(Terminal): iosevka superiority
Iosevka is better than berkeley mono. I like it sometimes, but it's too
playful. Iosevka is so straight and to the point, I like its no-nonsense
approach. It's also very good for coding specifically because it is so
skinny, allowing even more text to fit on the page (which is important
in things like lisps that have crazy indentation defaults)